♻️ refactor: entities, shared 폴더 구조 수정, ui 합성 컴포넌트로 수정
