+++ src/entities/model/comment.ts
@@ -1,0 +1,10 @@
+export interface Comment {
+  id: number
+  body: string
+  postId: number
+  userId: number
+  likes: number
+  user: {
+    username: string
+  }
+}
