+++ src/entities/model/index.ts
@@ -1,0 +1,3 @@
+export * from './post'
+export * from './user'
+export * from './comment'
