+++ src/entities/model/post.ts
@@ -1,0 +1,21 @@
+import { User } from './user'
+
+export interface Post {
+  id: number
+  title: string
+  body: string
+  userId: number
+  tags?: string[]
+  author?: User
+  reactions?: {
+    likes: number
+    dislikes: number
+  }
+}
+
+export interface PostsResponse {
+  posts: Post[]
+  total: number
+  skip: number
+  limit: number
+}
